pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowscale-nodes"
version = "0.3.16"
description = "FlowScale Nodes for ComfyUI - Enhanced media handling and utility nodes"
readme = "README.md"
authors = [
    { name = "w3rc" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "aiofiles",
    "aiohttp",
    "boto3",
    "httpx",
    "python-dotenv",
    "GitPython",
    "botocore",
    "pillow>=10.1.0",
    "pillow-avif-plugin>=1.5.0",
    "pillow-heif>=1.0.0",
    "numpy",
    "PyPDF2",
    "opencv-python",
    "soundfile",
    "mutagen",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pre-commit",
]

[project.urls]
Homepage = "https://github.com/flowscale-ai/flowscale-nodes"
Documentation = "https://github.com/flowscale-ai/flowscale-nodes/blob/main/README.md"

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
    "pre-commit",
]

[tool.hatch.build.targets.wheel]
packages = ["api", "nodes", "utilitynodes", "web"]

[tool.ruff]
# Assume Python 3.9+
target-version = "py39"
# Same as Black.
line-length = 100
indent-width = 4

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",    # isort
    "C90",  # mccabe complexity
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex (we'll handle this case by case)
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 10.
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["flowscale_nodes"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3